Refactor code to prepare to implement GitHub searching
