+++ src/gsc/__init__.py
@@ -1,10 +1,10 @@
-from gsc.command_line import cli, gitlab_cli
+from gsc.command_line import cli
 from gsc.di.application_container import ApplicationContainer
 
 
 def main():
     app_container = ApplicationContainer()
-    app_container.wire(modules=[gitlab_cli])
+    app_container.wire(packages=["gsc"])
     cli.app()
