+++ src/gsc/command_line/gitlab_cli.py
@@ -2,85 +2,36 @@ import os.path
 import click
 from dependency_injector.wiring import Provide, inject
 from gsc.di.application_container import ApplicationContainer
-from gsc.utils import is_valid_environment_name
+from gsc.constants import GitLabConstant
+from gsc.command_line.env_cli import environment
 from gsc.command_line import keep_main_thread_running
-from gsc.command_line.observer.gitlab_observer import (
+from gsc.observer.gitlab_observer import (
     GitLabParam,
     GitLabPrintObserver,
-    GitLabExportObserver,
 )
-from gsc.config import AppConfig, Env, EnvConfig, GitLabConfig
+from gsc.config import AppConfig, GitLabConfig
+from gsc.observer.plugin import ExportPlugin
 from gsc.use_cases.gitlab_search_use_case import (
     GitLabSearchGroupUseCase,
     GitLabSearchProjectUseCase,
 )
 
 
-@click.group("gl", help="Search in GitLab repositories.")
-def gitlab_cli():
-    pass
+@click.group("gl", help=f"Search in {GitLabConstant.NAME} repositories.")
+@click.pass_context
+@inject
+def gitlab_cli(
+    ctx, config: GitLabConfig = Provide[ApplicationContainer.gitlab_module.config]
+):
+    ctx.obj = config
+
+
+gitlab_cli.add_command(environment)
 
 
 @gitlab_cli.command(
-    "env", help="Setup the environment for searching, support multiple environments."
+    "search", help=f"Search the content in {GitLabConstant.NAME} repositories."
 )
-@click.option(
-    "-n",
-    "--new",
-    type=str,
-    metavar="<environment>",
-    help="Create new or override the env information if it was existed.",
-)
-@click.option(
-    "-d",
-    "--default",
-    type=str,
-    metavar="<environment>",
-    help="Set the environment as default.",
-)
-@click.option(
-    "-r",
-    "--remove",
-    type=str,
-    metavar="<environment>",
-    help="Remove the environment.",
-)
-@click.option(
-    "-i",
-    "--info",
-    type=str,
-    metavar="<environment>",
-    help="Display the environment info.",
-)
-@click.option(
-    "-l",
-    "--list",
-    "show_list",
-    is_flag=True,
-    default=False,
-    help="List all environment.",
-)
-@click.pass_context
-@inject
-def environment(
-    ctx, config: GitLabConfig = Provide[ApplicationContainer.gitlab.config], **kwargs
-):
-
-    if kwargs.get("show_list"):
-        __show_list_envs(config)
-    elif kwargs.get("new"):
-        __create_new_environment(config, kwargs.get("new"))
-    elif kwargs.get("default"):
-        __set_default_environment(config, kwargs.get("default"))
-    elif kwargs.get("remove"):
-        __remove_environment(config, kwargs.get("remove"))
-    elif kwargs.get("info"):
-        __show_detail_environment(config, kwargs.get("info"))
-    else:
-        click.secho(environment.get_help(ctx))
-
-
-@gitlab_cli.command("search", help="Search the content in GitLab repositories.")
 @click.argument("keyword", type=str, metavar="<keyword>")
 @click.option(
     "-g",
@@ -123,8 +74,8 @@ def environment(
 @inject
 def search(
     ctx,
-    app_config: AppConfig = Provide[ApplicationContainer.gitlab.app_config],
-    gitlab_config: AppConfig = Provide[ApplicationContainer.gitlab.config],
+    app_config: AppConfig = Provide[ApplicationContainer.gitlab_module.app_config],
+    gitlab_config: AppConfig = Provide[ApplicationContainer.gitlab_module.config],
     **kwargs,
 ):
     gitlab_config.set_session_env("")
@@ -168,7 +119,7 @@ def search(
 
 
 def __is_validate_output_path(path: str):
-    return os.path.splitext(path)[1] in GitLabExportObserver.MARKDOWN_EXTENSION
+    return os.path.splitext(path)[1] in ExportPlugin.EXTENSION_SUPPORTED
 
 
 @keep_main_thread_running
@@ -176,13 +127,11 @@ def __is_validate_output_path(path: str):
 def __search_in_group(
     param: GitLabParam,
     usecase: GitLabSearchGroupUseCase = Provide[
-        ApplicationContainer.gitlab.search_group_use_case
+        ApplicationContainer.gitlab_module.search_group_use_case
     ],
 ):
     param.is_search_group = True
     usecase.on_searching().subscribe(GitLabPrintObserver(param=param))
-    if param.output_path and not param.is_debug:
-        usecase.on_searching().subscribe(GitLabExportObserver(param=param))
     usecase.search(param.input_group, param.keyword)
 
 
@@ -191,63 +140,8 @@ def __search_in_group(
 def __search_in_project(
     param: GitLabParam,
     usecase: GitLabSearchProjectUseCase = Provide[
-        ApplicationContainer.gitlab.search_proj_use_case
+        ApplicationContainer.gitlab_module.search_proj_use_case
     ],
 ):
     usecase.on_searching().subscribe(GitLabPrintObserver(param=param))
-    if param.output_path and not param.is_debug:
-        usecase.on_searching().subscribe(GitLabExportObserver(param=param))
     usecase.search(param.input_project, param.keyword)
-
-
-def __show_list_envs(config: EnvConfig):
-    envs = config.get_all_envs()
-    if not envs:
-        click.secho("There is no environment.")
-        return
-
-    for env in envs:
-        name = env.name
-        if config.is_default_env(name):
-            click.secho(f"* {name}")
-        else:
-            click.secho(f"  {name}")
-
-
-def __create_new_environment(config: EnvConfig, name: str):
-    if not is_valid_environment_name(name):
-        click.secho(
-            "Environment name only contains letters, numbers, underscores, dashes..., not space"
-        )
-        return
-
-    new_env = Env(name=name)
-
-    host = click.prompt(f'Please input host name for "{name}" environment ', type=str)
-    if host:
-        new_env.host_name = host
-
-    token = click.prompt(f'Please input token for "{name}" environment ', type=str)
-    if token:
-        new_env.private_token = token
-
-    config.set_env(new_env)
-
-
-def __remove_environment(config: EnvConfig, env_name: str):
-    config.remove_env(env_name)
-
-
-def __show_detail_environment(config: EnvConfig, name: str):
-    if not config.is_env_existed(name):
-        click.secho(f'"{name}" environment is not found.')
-        return
-
-    environ = config.get_env(name)
-    click.secho(f'"{name}" environment :')
-    click.secho(f"Host name : {environ.host_name}")
-    click.secho(f"Private token : {environ.private_token}")
-
-
-def __set_default_environment(config: EnvConfig, env_name: str):
-    config.set_default_env(env_name)
